Пункты меню вынесены в константу
