+++ src/constants/navigation.ts
@@ -1,0 +1,19 @@
+export type NavItem = {
+    link: string,
+    path: string
+}
+
+export const NAVIGATION: NavItem[] = [
+    {
+        link: "Каталог",
+        path: "/",
+    },
+    {
+        link: "Сервис",
+        path: "/services",
+    },
+    {
+        link: "Контакты",
+        path: "/contacts",
+    },
+]
